LPS-65226 Move
