+++ modules/apps/foundation/portal-cache/portal-cache-single/src/main/java/com/liferay/portal/cache/single/internal/activator/GatekeeperActivator.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.scheduler.single.internal;
+package com.liferay.portal.cache.single.internal.activator;
 
+import com.liferay.portal.cache.single.internal.bootstrap.SinglePortalCacheBootstrapLoaderFactory;
+import com.liferay.portal.cache.single.internal.distribution.SinglePortalCacheReplicatorFactory;
 import com.liferay.portal.kernel.util.ReleaseInfo;
 
 import org.osgi.service.component.ComponentContext;
@@ -24,7 +26,7 @@ import org.osgi.service.component.annotations.Component;
  * @author Shuyang Zhou
  */
 @Component(immediate = true)
-public class GateKeeper {
+public class GatekeeperActivator {
 
 	@Activate
 	public void activate(ComponentContext componentContext) {
@@ -35,7 +37,9 @@ public class GateKeeper {
 		}
 
 		componentContext.enableComponent(
-			SingleSchedulerEngineConfigurator.class.getName());
+			SinglePortalCacheBootstrapLoaderFactory.class.getName());
+		componentContext.enableComponent(
+			SinglePortalCacheReplicatorFactory.class.getName());
 	}
 
 }
